% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drain.R
\name{drain}
\alias{drain}
\title{Clear package options}
\usage{
drain(slot_name)
}
\arguments{
\item{slot_name}{(optional) slot to clear from options()}
}
\description{
Clear options of previously specified content
}
