% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config}
\alias{read_config}
\title{Handle configuration data from a file}
\usage{
read_config(file)
}
\arguments{
\item{file}{string: path to file. Readable formats are \code{.yml} and \code{.json}.}
}
\value{
A \code{list} containing data from the specified file.
}
\description{
This is primarily an internal function for importing configuration
information from a file. It is called by \code{brew()}, and detects \code{.yml} or
\code{.json} files by their file extentions; all the actual work is done by
\code{yaml::read_yaml} and \code{jsonlite::read_json} respectively. It is available as
an exported function so that users can check their data is being imported
correctly, and for developers who may wish to intercept configuration files
for checking purposes.
}
