% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_config.R
\name{read_config_file}
\alias{read_config_file}
\title{Get config data from a file}
\usage{
read_config_file(file)
}
\arguments{
\item{file}{string: path to file. Readable formats are \code{.yml} and \code{.json}.}
}
\description{
In test. Looks for a file in the working directory
}
