% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{brew}
\alias{brew}
\alias{brew_package}
\alias{brew_interactive}
\title{Set up potions for easy data retrieval}
\usage{
brew(.data, .slot, .pkg)

brew_package(.data, .pkg)

brew_interactive(.data, .slot)
}
\arguments{
\item{.data}{a list containing data to be stored via `options()`.}

\item{.slot}{string: optional name to mandate where data is stored. Defaults 
to a random string generated by `stringi::stri_rand_strings`.}

\item{.pkg}{string: package name that `potions` is being used within. If 
using {potions} within a package development process, set this argument 
instead of `slot`.}
}
\description{
Start-up function to place a list into `options` with a specified 
slot name.
}
\details{
The default method is to use `brew` and set either `pkg` or `slot`, but not
both. Alternatively you can use `brew_package()` or `brew_interactive()`. 
Note that if neither `.slot` or `.pkg` are set, `potions` defaults to `.slot`
, unless `.pkg` information has previously been supplied (and `.slot` 
information has not). This might be undesirable in a package development 
situation.

If no `.data` argument is given, this function sets up an empty `potions` 
object in `options("potions-pkg")`. See `potions-class` for more information.

If the user repeatly calls `brew()`, later list entries overwrite early 
entries. Whole lists are not overwritten unless all top-level entry names 
match.
}
