% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brew.R
\name{brew}
\alias{brew}
\title{Set up potions for easy data retrieval}
\usage{
brew(slot_name, x)
}
\arguments{
\item{slot_name}{string: where should data be stored?}

\item{x}{a list containing data required at later stages}
}
\description{
Start-up function to place a list into `options` with a specified 
slot name.
}
\details{
If used successively, later list entries overwrite early entries. The whole
list is not overwritten.
}
