% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potions_class.R, R/print.R
\name{potions-class}
\alias{potions-class}
\alias{create_potions}
\alias{explore_potions}
\alias{print.potions}
\title{Class for `potions` lists}
\usage{
create_potions()

explore_potions()

\method{print}{potions}(x, ...)
}
\description{
Early days

Similar to `pour`, but returns all data, while `pour` only ever returns
data for a single slot or package
}
