% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potions_class.R, R/print.R
\name{potions-class}
\alias{potions-class}
\alias{create_potions}
\alias{print.potions}
\title{Methods for `potions` data}
\usage{
create_potions()

\method{print}{potions}(x, ...)
}
\arguments{
\item{x}{An object of class `potions`}

\item{...}{Any further arguments to `print()`}
}
\description{
Data storage for information passed to `{potions}`.
}
