% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potions_class.R, R/print.R
\name{potions-class}
\alias{potions-class}
\alias{create_potions}
\alias{print.potions}
\title{Methods for \code{potions} data}
\usage{
create_potions()

\method{print}{potions}(x, ...)
}
\arguments{
\item{x}{An object of class \code{potions}}

\item{...}{Any further arguments to \code{print()}}
}
\value{
In the case of \code{create_potions()}, an empty \code{potions} object.
\code{print.potions()} displays a \code{potions} object using \code{lobstr::tree()}.
}
\description{
This package stores data in a list-like format, named class \code{potions}. It
contains three entries: \code{slots} contains data stored in 'interactive' mode;
\code{packages} contains data from packages built using \code{potions}; and \code{mapping}
stores data to understand the contents of the other two slots.
}
