% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{pour}
\alias{pour}
\title{Retrieve a specified slot from options}
\usage{
pour(..., slot_name)
}
\arguments{
\item{...}{strings: what slots should be returned}

\item{slot_name}{Optional manual override to default slot. Useful for
ensuring no clashes when used within a package}
}
\value{
a vector
}
\description{
UI based on here; i.e. uses strings separated by commas
}
