% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pour.R
\name{pour}
\alias{pour}
\title{Retrieve a specified slot from options}
\usage{
pour(...)
}
\arguments{
\item{...}{strings: what slots should be returned}
}
\description{
UI based on here; i.e. uses strings separated by commas
}
